+++ src/link.ts
@@ -14,7 +14,7 @@ type LinkInit = {
 /**
  * The Link object represents a hyperlink.
  */
-export default class Link {
+export class Link {
 
   /**
    * The base href of the parent document. Used for expanding relative links.
@@ -96,3 +96,16 @@ export default class Link {
   }
 
 }
+
+export default Link;
+
+/**
+ * A LinkSet is just a map of links, indexes by their rel
+ */
+export type LinkSet = Map<string, Link[]>;
+
+/**
+ * The LinkNotFound error gets thrown whenever something tries to follow a
+ * link by its rel, that doesn't exist
+ */
+export class LinkNotFound extends Error {}